% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dog_walk.R
\name{dog_walk}
\alias{dog_walk}
\title{A function that reports back the total number of dog walks and the location}
\usage{
dog_walk(total_walks, walk_location, day)
}
\arguments{
\item{total_walks}{total number of dog walks for the give day}

\item{walk_location}{wherre the walk took place}

\item{day}{what day the walk took place}
}
\description{
A function that reports back the total number of dog walks and the location
}
\examples{
dog_walk(3, "the beach", "friday")

}
